Fixed yarn ball recipe not working due to wrong workbench type. Added `/cat despawn` and `/cat spawn` commands for better cat management. Added CatsManager as centralized data access layer for all cat operations. Added owner name persistence (stored alongside owner UUID in cat data). Added persistent cat data storage in CatsDataResource (saved to `CatsData.json`). Refactored all commands to use CatsManager instead of direct resource access. Refactored cat name storage to use Nameplate component instead of CatOwnerComponent. Improved data architecture with strict command → manager → resource flow. Optimized CatDataEntry codec with named constants and streamlined implementation.

+++ src/main/java/de/markusbordihn/cats/inventory/InventoryHelper.java
@@ -42,7 +42,7 @@ public class InventoryHelper {
     }
 
     byte activeSlot = inventory.getActiveHotbarSlot();
-    inventory.getHotbar().removeItemStackFromSlot((short) activeSlot, 1);
+    inventory.getHotbar().removeItemStackFromSlot(activeSlot, 1);
     LOGGER.at(Level.FINE).log(
         "Consumed 1x %s from player inventory at slot %d", itemName, activeSlot);
   }
